backend/requirements.txt
streamlit
requests
Pillow
fastapi
uvicorn
transformers
torch 
python-multipart
agno
ollama
groq
python-dotenv
duckduckgo-search
streamlit
ddgs
groq
langchain
langchain_community
langchain_pinecone
pydub
tiktoken
sentence-transformers
pinecone-text
langchain_groq
streamlit
moviepy==1.0.3
langchain-core 
gTTS
sqlalchemy
bcrypt==4.0.1 
passlib==1.7.4
python-multipart 
authlib 
itsdangerous
starlette
